requirements.txt
flask
pymongo
pymongo[srv]
python-dotenv